# add function to run VM and return its final state