Function `debug_terminal_state` is useful for testing snippets. Notably, it requires considerable less RAM than `debug` since no intermediate states are recorded.