@Sword-Smith 2023-06-20
I measured running time for a few snippets that read from secret-in:

**Before this commit**
PASS [   2.956s] tasm-lib mmr::verify_from_secret_in::mmr_verify_from_secret_in_tests::mmra_ap_verify_test_pbt
PASS [   2.436s] tasm-lib mmr::load_from_secret_in_then_verify::tests::load_from_secret_in_then_verify_test_safe_lists

**After this commit**

 PASS [   0.827s] tasm-lib mmr::load_from_secret_in_then_verify::tests::load_from_secret_in_then_verify_test_safe_lists
 PASS [   1.138s] tasm-lib mmr::verify_from_secret_in::tests::mmra_ap_verify_test_pbt
 
 So more than a 2x speedup from merging this PR. Almost certainly through the use of `VecDeque` instead of `Vec` for standard-in and secret-in.